+++ R/deprec-failwith.r
@@ -1,20 +1,14 @@
 #' Fail with specified value.
 #'
-#' Deprecated. Please use [purrr::possibly()] instead.
+#' \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
+#' Please use [purrr::possibly()] instead.
 #'
 #' @param default default value
 #' @param f function
 #' @param quiet all error messages be suppressed?
 #' @return a function
-#' @seealso [plyr::try_default()]
 #' @keywords internal
 #' @export
 failwith <- function(default = NULL, f, quiet = FALSE) {
-  warn("Deprecated: please use `purrr::possibly()` instead")
-
-  function(...) {
-    out <- default
-    try(out <- f(...), silent = quiet)
-    out
-  }
+  lifecycle::deprecate_stop("0.7.0", "failwith()", "purrr::possibly()")
 }